requirements.txt
langchain==0.0.136
openai
faiss-cpu
streamlit
streamlit-chat
